+++ src/narrativepipeline/NarrativePipeline.py
@@ -1,5 +1,6 @@
-import os, json
 from typing import List
+import os, json, sys, traceback
+
 
 from torch.utils.data import DataLoader
 import torch.nn as torch_nn
@@ -7,19 +8,19 @@ import torch
 from transformers import AdamW
 
 
-from modules.narrativepipeline.utils import CustomDataset, build_vocab, Vocab
-from modules.ans_infer.Transformer import TransDecoder
-from modules.utils import check_exist
-from modules.Reasoning.GraphReasoning import GraphReasoning
-from modules.finegrained.BertEmbedding import SimpleBertEmbd
+from src.narrativepipeline.utils import CustomDataset, build_vocab, Vocab
+from src.ans_infer.Transformer import TransDecoder
+from src.utils import check_exist
+from src.reasoning.MemoryGraph import GraphBasedMemoryReasoning
+from src.finegrained.FineGrain import FineGrain
 from configs import args, logging, PATH
 
 class  NarrativePipeline(torch_nn.Module):
     def __init__(self, vocab):
         super().__init__()
 
-        self.embd_layer = SimpleBertEmbd()
-        self.reasoning  = GraphReasoning()
+        self.embd_layer = FineGrain()
+        self.reasoning  = GraphBasedMemoryReasoning()
         self.ans_infer  = TransDecoder(vocab, self.embd_layer.embedding)
 
     def forward(self, ques, ques_mask, ans, ans_mask,
@@ -35,43 +36,28 @@ class  NarrativePipeline(torch_nn.Module):
         # edge_indx      : [b, 2, n_edges]
         # edge_len       : [b]
 
-        batch   = ques.shape[0]
 
         ####################
         # Embed question, paras and answer
-        # with BERT
         ####################
-        ques_embd, para_embd, ans_seq_embd = \
-            self.embd_layer(ques, ques_mask, paras, paras_mask, ans, ans_mask)
-        # ques_embd     : [b, 768]
-        # paras_seq_embd: [b, n_paras, 768]
-        # ans_seq_embd  : [b, seq_len_ans, 768] or None
+        ques, paras, ans = self.embd_layer(ques, ques_mask, paras, paras_mask, ans, ans_mask)
+        # ques : [b, seq_len_ques, d_hid]
+        # paras: [b, n_paras, seq_len_ques, d_hid]
+        # ans  : [b, seq_len_ans, d_hid] or None
 
-        ####################
-        # Do reasoning with IAL
-        ####################
-
-        # Add question to paras to create tensor 'node_feat'
-        node_feat   = torch.cat((ques_embd.unsqueeze(1), para_embd), dim=1).to(args.device)
-        # node_feat : [b, 1+n_paras, 768]
-
-        # Add 1 to each of tensor 'paras_len' because question is added to paras
-        paras_len   = paras_len + torch.ones((batch,)).to(args.device)
-
-        Y       = self.reasoning(node_feat, edge_indx, paras_len.int(), edge_len)
-        # Y: [b, d_hid=256]
 
         ####################
-        # Generate answer with PGD
+        # Do reasoning
         ####################
-        # Repeat tensor 'Y' to dim [b, seq_len_para * n_paras, d_hid*2]
-        d       = args.seq_len_para * args.n_paras
-        Y       = Y.unsqueeze(1).repeat(1, d, 2)
-        # [b, seq_len_para * n_paras, d_hid*2]
+        Y       = self.reasoning(ques, paras)
+        # [b, n_nodes + n_paras*seq_len_para, d_hid]
 
-        pred    = self.ans_infer(Y, ans_seq_embd,
-                                 ans_mask, is_inferring)
-        # pred: [batch, max_len_ans, d_vocab + seq_len_cntx]
+        ####################
+        # Generate answer
+        ####################
+
+        pred    = self.ans_infer(Y, ans, is_inferring)
+        # [b, seq_len_ans, d_vocab]
 
         return pred
 
@@ -152,34 +138,49 @@ class Trainer():
             iterator_train  = DataLoader(dataset_train, batch_size=args.batch, shuffle=True)
 
             for batch in iterator_train:
-                ques            = batch['ques'].to(args.device)
-                ques_mask       = batch['ques_mask'].to(args.device)
-                ans1_bert_ids   = batch['ans1_bert_ids'].to(args.device)
-                ans1_bert_mask  = batch['ans1_bert_mask'].to(args.device)
-                ans1_vocab_ids  = batch['ans1_vocab_ids'].to(args.device)
-                ans2_vocab_ids  = batch['ans2_vocab_ids'].to(args.device)
-                paras           = batch['paras'].to(args.device)
-                paras_len       = batch['paras_len'].to(args.device)
-                paras_mask      = batch['paras_mask'].to(args.device)
-                edge_indx       = batch['edge_indx'].to(args.device)
-                edge_len        = batch['edge_len'].to(args.device)
+                pred_result = []
+
+                ques        = batch['ques'].to(args.device).float()
+                ques_mask   = batch['ques_mask'].to(args.device)
+                ans1        = batch['ans1'].to(args.device).float()
+                ans1_mask   = batch['ans1_mask'].to(args.device)
+                ans1_ids    = batch['ans1_ids'].to(args.device)
+                ans2_ids    = batch['ans2_ids'].to(args.device)
+                paras       = batch['paras'].to(args.device).float()
+                paras_mask  = batch['paras_mask'].to(args.device)
 
 
                 optimizer.zero_grad()
 
-                pred        = model(ques, ques_mask, ans1_bert_ids, ans1_bert_mask,
-                                    paras, paras_len, paras_mask, edge_indx,
-                                    edge_len, is_inferring=False)
+                pred        = model(ques, ques_mask, ans1, ans1_mask,
+                                    paras, paras_mask, is_inferring=False)
                 # pred: [batch, seq_len_ans, d_vocab]
+
+
+                # This piece of code is to show tokens of pred during training
+                _, indices = torch.softmax(pred, dim=2).topk(k=1, dim=2)
+                indices = indices.squeeze(2)
+                for pred_, ans1_, ans2_ in zip(indices, ans1_ids, ans2_ids):
+                    pred_result.append({
+                        'pred': ' '.join(self.vocab.itos(pred_.tolist())),
+                        'ans1': ' '.join(self.vocab.itos(ans1_.tolist())),
+                        'ans2': ' '.join(self.vocab.itos(ans2_.tolist()))
+                    })
+
+                with open("backup/pred_train.json", 'a+') as result_file:
+                    json.dump(pred_result, result_file, indent=2, ensure_ascii=False)
+
+
+
+                # pred: [b, seq_len_ans, d_vocab]
                 pred_flat   = pred.view(-1, args.d_vocab)
-                ans1_flat   = ans1_vocab_ids.view(-1)
-                ans2_flat   = ans2_vocab_ids.view(-1)
+                ans1_flat   = ans1_ids.view(-1)
+                ans2_flat   = ans2_ids.view(-1)
 
                 loss        = 0.7 * criterion(pred_flat, ans1_flat) +\
                               0.3 * criterion(pred_flat, ans2_flat)
-
-
                 loss.backward()
+
                 torch_nn.utils.clip_grad_value_(model.parameters(), clip_value=1)
                 optimizer.step()
                 scheduler.step()
@@ -200,41 +201,37 @@ class Trainer():
 
         with torch.no_grad():
             for eval_file in dataset_test.file_names:
-                logging.info(f"Eval with file: {eval_file}")
+                logging.info(f"Test with file: {eval_file}")
 
                 dataset_test.read_shard(eval_file)
                 iterator_test  = DataLoader(dataset_test, batch_size=args.batch)
                 n_samples   += len(dataset_test)
 
                 for batch in iterator_test:
-                    ques            = batch['ques'].to(args.device)
-                    ques_mask       = batch['ques_mask'].to(args.device)
-                    ans1_bert_ids   = batch['ans1_bert_ids'].to(args.device)
-                    ans1_bert_mask  = batch['ans1_bert_mask'].to(args.device)
-                    ans1_vocab_ids  = batch['ans1_vocab_ids'].to(args.device)
-                    ans2_vocab_ids  = batch['ans2_vocab_ids'].to(args.device)
-                    paras           = batch['paras'].to(args.device)
-                    paras_len       = batch['paras_len'].to(args.device)
-                    paras_mask      = batch['paras_mask'].to(args.device)
-                    edge_indx       = batch['edge_indx'].to(args.device)
-                    edge_len        = batch['edge_len'].to(args.device)
+                    ques        = batch['ques'].to(args.device)
+                    ques_mask   = batch['ques_mask'].to(args.device)
+                    ans1        = batch['ans1'].to(args.device)
+                    ans1_mask   = batch['ans1_mask'].to(args.device)
+                    ans1_ids    = batch['ans1_ids'].to(args.device)
+                    ans2_ids    = batch['ans2_ids'].to(args.device)
+                    paras       = batch['paras'].to(args.device)
+                    paras_mask  = batch['paras_mask'].to(args.device)
 
-                    pred            = model(ques, ques_mask, ans1_bert_ids, ans1_bert_mask,
-                                            paras, paras_len, paras_mask, edge_indx,
-                                            edge_len, is_inferring=False)
+                    pred        = model(ques, ques_mask, ans1, ans1_mask,
+                                        paras, paras_mask, is_inferring=False)
                     # pred: [batch, seq_len_ans, d_vocab]
                     pred_flat   = pred.view(-1, args.d_vocab)
-                    ans1_flat   = ans1_vocab_ids.view(-1)
-                    ans2_flat   = ans2_vocab_ids.view(-1)
+                    ans1_flat   = ans1_ids.view(-1)
+                    ans2_flat   = ans2_ids.view(-1)
 
                     loss        = 0.7 * criterion(pred_flat, ans1_flat) +\
-                                0.3 * criterion(pred_flat, ans2_flat)
+                                  0.3 * criterion(pred_flat, ans2_flat)
 
-                    loss_test += loss.detach().item()
+                    loss_test   += loss.detach().item()
+
 
                     logging.info(f"  test: batch {nth_batch:4d} | loss: {loss:8.6f}")
                     nth_batch += 1
-
 
         return loss_test / nth_batch
 
@@ -288,6 +285,9 @@ class Trainer():
 
             logging.info("Switch Answerability.")
             dataset_train.switch_answerability()
+
+            model.reasoning.save_memory()
+
             if loss_test < best_loss_test:
                 best_loss_test = loss_test
                 self.save_model(model)
@@ -308,22 +308,22 @@ class Trainer():
         ###############################
         # Load data
         ###############################
-        dataset_valid   = CustomDataset(os.path.dirname(PATH['dataset_para']).replace("[SPLIT]", "validation"))
+        dataset_valid   = CustomDataset(os.path.dirname(PATH['dataset']).replace("[SPLIT]", "valid"),
+                                      self.vocab)
 
         ###############################
         # Defind model and associated stuffs
         ###############################
-        model       = NarrativePipeline(self.vocab).to(args.device)
-
+        model       = NarrativePipeline(self.vocab).to(args.device).to(args.device).float()
         model       = self.load_model(model)
 
         ###############################
         # Start infering
         ###############################
-
         pred_result = []
 
         model.eval()
+
         with torch.no_grad():
             for valid_file in dataset_valid.file_names:
                 logging.info(f"Valid with file: {valid_file}")
@@ -332,23 +332,18 @@ class Trainer():
 
                 iterator_valid  = DataLoader(dataset_valid, batch_size=args.batch)
                 for batch in iterator_valid:
-                    ques            = batch['ques'].to(args.device)
-                    ques_mask       = batch['ques_mask'].to(args.device)
-                    ans1_bert_ids   = batch['ans1_bert_ids'].to(args.device)
-                    ans1_bert_mask  = batch['ans1_bert_mask'].to(args.device)
-                    ans1_vocab_ids  = batch['ans1_vocab_ids'].to(args.device)
-                    ans2_vocab_ids  = batch['ans2_vocab_ids'].to(args.device)
-                    paras           = batch['paras'].to(args.device)
-                    paras_len       = batch['paras_len'].to(args.device)
-                    paras_mask      = batch['paras_mask'].to(args.device)
-                    edge_indx       = batch['edge_indx'].to(args.device)
-                    edge_len        = batch['edge_len'].to(args.device)
+                    ques        = batch['ques'].to(args.device)
+                    ques_mask   = batch['ques_mask'].to(args.device)
+                    ans1        = batch['ans1'].to(args.device)
+                    ans1_mask   = batch['ans1_mask'].to(args.device)
+                    ans1_ids    = batch['ans1_ids'].to(args.device)
+                    ans2_ids    = batch['ans2_ids'].to(args.device)
+                    paras       = batch['paras'].to(args.device)
+                    paras_mask  = batch['paras_mask'].to(args.device)
 
-                    pred            = model(ques, ques_mask, None, None,
-                                            paras, paras_len, paras_mask, edge_indx,
-                                            edge_len, is_inferring=True)
-
-                    for pred_, ans1_, ans2_ in zip(pred, ans1_vocab_ids, ans2_vocab_ids):
+                    pred        = model(ques, ques_mask, ans1, ans1_mask,
+                                        paras, paras_mask, is_inferring=True)
+                    for pred_, ans1_, ans2_ in zip(pred, ans1_ids, ans2_ids):
                         pred_result.append({
                             'pred': ' '.join(self.vocab.itos(pred_.tolist())),
                             'ans1': ' '.join(self.vocab.itos(ans1_.tolist())),
@@ -358,16 +353,21 @@ class Trainer():
                     with open(PATH['prediction'], 'a+') as result_file:
                         json.dump(pred_result, result_file, indent=2, ensure_ascii=False)
 
-
-        with open(PATH['prediction'], 'w+') as result_file:
-            json.dump(pred_result, result_file, indent=2, ensure_ascii=False)
-
 if __name__ == '__main__':
     logging.info("* Start NarrativePipeline")
 
+    try:
+        narrative_pipeline  = Trainer()
 
-    narrative_pipeline  = Trainer()
+        if args.task == "train":
+            narrative_pipeline.trigger_train()
+        elif args.task == "infer":
+            narrative_pipeline.trigger_infer()
+        else:
+            raise AssertionError("'task' argument incorrect.")
 
-    # narrative_pipeline.trigger_train()
-
-    narrative_pipeline.trigger_infer()
+    except Exception as err:
+        exc_info = sys.exc_info()
+        with open(PATH['log'], "a+") as d_file:
+            traceback.print_exception(*exc_info, file=d_file)
+            sys.exit()